@@ -3,46 +3,46 @@
     [
       {
         "name": "webpack-sources",
-        "raw": "webpack-sources@^0.2.3",
-        "rawSpec": "^0.2.3",
+        "raw": "webpack-sources@^1.0.1",
+        "rawSpec": "^1.0.1",
         "scope": null,
-        "spec": ">=0.2.3 <0.3.0",
+        "spec": ">=1.0.1 <2.0.0",
         "type": "range"
       },
       "/tmp/build/88bfaeb5/gpcc/frontend/node_modules/webpack"
     ]
   ],
-  "_from": "webpack-sources@>=0.2.3 <0.3.0",
-  "_id": "webpack-sources@0.2.3",
+  "_from": "webpack-sources@>=1.0.1 <2.0.0",
+  "_id": "webpack-sources@1.0.1",
   "_inCache": true,
   "_installable": true,
   "_location": "/webpack-sources",
-  "_nodeVersion": "7.4.0",
+  "_nodeVersion": "8.0.0",
   "_npmOperationalInternal": {
-    "host": "packages-18-east.internal.npmjs.com",
-    "tmp": "tmp/webpack-sources-0.2.3.tgz_1490398694081_0.5428378875367343"
+    "host": "s3://npm-registry-packages",
+    "tmp": "tmp/webpack-sources-1.0.1.tgz_1496500323403_0.9926682878285646"
   },
   "_npmUser": {
     "email": "tobias.koppers@googlemail.com",
     "name": "sokra"
   },
-  "_npmVersion": "4.0.5",
+  "_npmVersion": "5.0.0",
   "_phantomChildren": {},
   "_requested": {
     "name": "webpack-sources",
-    "raw": "webpack-sources@^0.2.3",
-    "rawSpec": "^0.2.3",
+    "raw": "webpack-sources@^1.0.1",
+    "rawSpec": "^1.0.1",
     "scope": null,
-    "spec": ">=0.2.3 <0.3.0",
+    "spec": ">=1.0.1 <2.0.0",
     "type": "range"
   },
   "_requiredBy": [
     "/webpack"
   ],
-  "_resolved": "https://registry.npmjs.org/webpack-sources/-/webpack-sources-0.2.3.tgz",
-  "_shasum": "17c62bfaf13c707f9d02c479e0dcdde8380697fb",
+  "_resolved": "https://registry.npmjs.org/webpack-sources/-/webpack-sources-1.0.1.tgz",
+  "_shasum": "c7356436a4d13123be2e2426a05d1dad9cbe65cf",
   "_shrinkwrap": null,
-  "_spec": "webpack-sources@^0.2.3",
+  "_spec": "webpack-sources@^1.0.1",
   "_where": "/tmp/build/88bfaeb5/gpcc/frontend/node_modules/webpack",
   "author": {
     "name": "Tobias Koppers @sokra"
@@ -51,7 +51,7 @@
     "url": "https://github.com/webpack/webpack-sources/issues"
   },
   "dependencies": {
-    "source-list-map": "^1.1.1",
+    "source-list-map": "^2.0.0",
     "source-map": "~0.5.3"
   },
   "description": "Source code handling classes for webpack",
@@ -59,22 +59,23 @@
     "beautify-lint": "^1.0.3",
     "codecov.io": "^0.1.6",
     "coveralls": "^2.11.6",
-    "eslint": "^1.1.0",
+    "eslint": "^3.19.0",
     "eslint-plugin-nodeca": "^1.0.3",
     "istanbul": "^0.4.1",
     "js-beautify": "^1.5.10",
-    "mocha": "^2.3.4",
-    "should": "^8.0.2"
+    "mocha": "^3.4.2",
+    "should": "^11.2.1"
   },
   "directories": {},
   "dist": {
-    "shasum": "17c62bfaf13c707f9d02c479e0dcdde8380697fb",
-    "tarball": "https://registry.npmjs.org/webpack-sources/-/webpack-sources-0.2.3.tgz"
+    "integrity": "sha512-05tMxipUCwHqYaVS8xc7sYPTly8PzXayRCB4dTxLhWTqlKUiwH6ezmEe0OSreL1c30LAuA3Zqmc+uEBUGFJDjw==",
+    "shasum": "c7356436a4d13123be2e2426a05d1dad9cbe65cf",
+    "tarball": "https://registry.npmjs.org/webpack-sources/-/webpack-sources-1.0.1.tgz"
   },
   "files": [
     "lib/"
   ],
-  "gitHead": "828a85acb96faf69c1ba0b78ba6d3a2acc9dcd92",
+  "gitHead": "5cadaf4ab143a32cf994365db4b2ddacd0ab0eb9",
   "homepage": "https://github.com/webpack/webpack-sources#readme",
   "keywords": [
     "webpack",
@@ -122,5 +123,5 @@
     "test": "mocha --full-trace --check-leaks",
     "travis": "npm run cover -- --report lcovonly"
   },
-  "version": "0.2.3"
+  "version": "1.0.1"
 }
