@@ -1,3 +1,6 @@
+'use strict';
+/* global describe it */
+
 var assert = require('assert');
 var hash = require('../');
 
@@ -12,9 +15,9 @@ describe('Hash', function() {
       assert.equal(dgst, res);
 
       // Split message
-      var dgst = fn().update(msg.slice(0, 2), enc)
-                     .update(msg.slice(2), enc)
-                     .digest('hex');
+      dgst = fn().update(msg.slice(0, 2), enc)
+        .update(msg.slice(2), enc)
+        .digest('hex');
       assert.equal(dgst, res);
     }
   }
@@ -92,10 +95,11 @@ describe('Hash', function() {
         'ddaf35a193617abacc417349ae20413112e6fa4e89a97ea20a9eeee64b55d39a' +
             '2192992a274fc1a836ba3c23a3feebbd454d4423643ce80e2a9ac94fa54ca49f'
       ],
-      [ 'abcdefghbcdefghicdefghijdefghijkefghijklfghijklmghijklmn' +
-            'hijklmnoijklmnopjklmnopqklmnopqrlmnopqrsmnopqrstnopqrstu',
+      [
+        'abcdefghbcdefghicdefghijdefghijkefghijklfghijklmghijklmn' +
+          'hijklmnoijklmnopjklmnopqklmnopqrlmnopqrsmnopqrstnopqrstu',
         '8e959b75dae313da8cf4f72814fc143f8f7779c6eb9f7fa17299aeadb6889018' +
-            '501d289e4900f7e4331b99dec4b5433ac7d329eeb6dd26545e96e55b874be909'
+          '501d289e4900f7e4331b99dec4b5433ac7d329eeb6dd26545e96e55b874be909'
       ]
     ]);
   });
@@ -109,10 +113,11 @@ describe('Hash', function() {
         'cb00753f45a35e8bb5a03d699ac65007272c32ab0eded1631a8b605a43ff5bed' +
             '8086072ba1e7cc2358baeca134c825a7'
       ],
-      [ 'abcdefghbcdefghicdefghijdefghijkefghijklfghijklmghijklmn' +
-            'hijklmnoijklmnopjklmnopqklmnopqrlmnopqrsmnopqrstnopqrstu',
+      [
+        'abcdefghbcdefghicdefghijdefghijkefghijklfghijklmghijklmn' +
+          'hijklmnoijklmnopjklmnopqklmnopqrlmnopqrsmnopqrstnopqrstu',
         '09330c33f71147e83d192fc782cd1b4753111b173b3b05d22fa08086e3b0f712' +
-            'fcc7c71a557e2db966c3e9fa91746039'
+          'fcc7c71a557e2db966c3e9fa91746039'
       ]
     ]);
   });
