@@ -5,6 +5,7 @@
 "use strict";
 
 const path = require("path");
+const crypto = require("crypto");
 const RequestShortener = require("./RequestShortener");
 const ConcatSource = require("webpack-sources").ConcatSource;
 const RawSource = require("webpack-sources").RawSource;
@@ -131,6 +132,7 @@ class SourceMapDevToolPlugin {
 					if(currentSourceMappingURLComment !== false && /\.css($|\?)/i.test(file)) {
 						currentSourceMappingURLComment = currentSourceMappingURLComment.replace(/^\n\/\/(.*)$/, "\n/*$1*/");
 					}
+					const sourceMapString = JSON.stringify(sourceMap);
 					if(sourceMapFilename) {
 						let filename = file;
 						let query = "";
@@ -139,22 +141,25 @@ class SourceMapDevToolPlugin {
 							query = filename.substr(idx);
 							filename = filename.substr(0, idx);
 						}
-						const sourceMapFile = compilation.getPath(sourceMapFilename, {
+						let sourceMapFile = compilation.getPath(sourceMapFilename, {
 							chunk,
 							filename,
 							query,
 							basename: basename(filename)
 						});
+						if(sourceMapFile.indexOf("[contenthash]") !== -1) {
+							sourceMapFile = sourceMapFile.replace(/\[contenthash\]/g, crypto.createHash("md5").update(sourceMapString).digest("hex"));
+						}
 						const sourceMapUrl = path.relative(path.dirname(file), sourceMapFile).replace(/\\/g, "/");
 						if(currentSourceMappingURLComment !== false) {
 							asset.__SourceMapDevToolData[file] = compilation.assets[file] = new ConcatSource(new RawSource(source), currentSourceMappingURLComment.replace(/\[url\]/g, sourceMapUrl));
 						}
-						asset.__SourceMapDevToolData[sourceMapFile] = compilation.assets[sourceMapFile] = new RawSource(JSON.stringify(sourceMap));
+						asset.__SourceMapDevToolData[sourceMapFile] = compilation.assets[sourceMapFile] = new RawSource(sourceMapString);
 						chunk.files.push(sourceMapFile);
 					} else {
 						asset.__SourceMapDevToolData[file] = compilation.assets[file] = new ConcatSource(new RawSource(source), currentSourceMappingURLComment
-							.replace(/\[map\]/g, () => JSON.stringify(sourceMap))
-							.replace(/\[url\]/g, () => `data:application/json;charset=utf-8;base64,${new Buffer(JSON.stringify(sourceMap), "utf-8").toString("base64")}`) // eslint-disable-line
+							.replace(/\[map\]/g, () => sourceMapString)
+							.replace(/\[url\]/g, () => `data:application/json;charset=utf-8;base64,${new Buffer(sourceMapString, "utf-8").toString("base64")}`) // eslint-disable-line
 						);
 					}
 				});
